Changing parsing output of biographical info
Info will now go into decl[u"biography"], whereas before it was
at the declaration level. This makes it easier to iterate over
declaration fields. It might break some things with the database
but we haven't used CouchDB for over a year and flat files fit
our workflow better. This is part of planned conversion to flat
file export.